+++ src/oemof/tabular/datapackage/reading.py
@@ -13,6 +13,7 @@ along with how to use the functions in this module.
 import collections.abc as cabc
 import json
 import re
+import typing
 import warnings
 from decimal import Decimal
 from itertools import chain, groupby, repeat
@@ -117,14 +118,14 @@ def deserialize_energy_system(cls, path, typemap={}, attributemap={}):
         FLOW_TYPE: HSN,
     }
 
-    for k, v in default_typemap.items():
-        typemap[k] = typemap.get(k, v)
+    for k, value in default_typemap.items():
+        typemap[k] = typemap.get(k, value)
 
     if attributemap.get(object) is None:
         attributemap[object] = {"name": "label"}
 
-    for k, v in attributemap.items():
-        if v.get("name") is None:
+    for k, value in attributemap.items():
+        if value.get("name") is None:
             attributemap[k]["name"] = "label"
 
     package = dp.Package(path)
@@ -133,8 +134,13 @@ def deserialize_energy_system(cls, path, typemap={}, attributemap={}):
     for r in package.resources:
         try:
             r.read()
-        except dp.exceptions.CastError:
-            raise dp.exceptions.CastError((cast_error_msg).format(r.name))
+        except dp.exceptions.CastError as e:
+            raise dp.exceptions.CastError(
+                "\n"
+                + (cast_error_msg).format(r.name)
+                + "\n"
+                + "\n ".join([str(i) for i in e.errors])
+            )
     empty = HSN()
     empty.read = lambda *xs, **ks: ()
     empty.headers = ()
@@ -304,13 +310,15 @@ def deserialize_energy_system(cls, path, typemap={}, attributemap={}):
 
     data["buses"] = {
         name: create(
-            mapping
-            if mapping
-            else raisestatement(
-                ValueError,
-                "Typemap is missing a mapping for '{}'.".format(
-                    bus.get("type", "bus")
-                ),
+            (
+                mapping
+                if mapping
+                else raisestatement(
+                    ValueError,
+                    "Typemap is missing a mapping for '{}'.".format(
+                        bus.get("type", "bus")
+                    ),
+                )
             ),
             {"label": name},
             bus["parameters"],
@@ -377,6 +385,159 @@ def deserialize_energy_system(cls, path, typemap={}, attributemap={}):
         for flow in (typemap.get(FLOW_TYPE, HSN),)
     }
 
+    period_data = {}
+    if package.get_resource("periods"):
+        df_periods = pd.DataFrame.from_dict(
+            package.get_resource("periods").read(keyed=True)
+        )
+        period_data["timeincrement"] = df_periods["timeincrement"].values
+        period_data["timeindex"] = pd.DatetimeIndex(df_periods["timeindex"])
+        period_data["periods"] = [
+            pd.DatetimeIndex(df["timeindex"])
+            for period, df in df_periods.groupby("periods")
+        ]
+        period_data["periods"] = [
+            pd.DatetimeIndex(i.values, freq=i.inferred_freq, name="timeindex")
+            for i in period_data["periods"]
+        ]
+        period_data["years"] = period_data["timeindex"].year.unique().values
+
+    def create_periodic_values(values, periods_index):
+        """
+        Create periodic values from given values and period_data.
+        The values are repeated for each period for the whole length e.g.
+        8760 values for hourly data in one period.
+
+        Parameters
+        ----------
+        values : list
+            List of values to be repeated.
+        periods_index : list
+            List containing periods datetimeindex.
+
+        Returns
+        -------
+        list
+            List of periodic values.
+        """
+        # check if length of list equals number of periods
+        if len(values) != len(periods_index):
+            raise ValueError(
+                "Length of values does not equal number of periods."
+            )
+
+        # create timeseries with periodic values
+        periodic_values = pd.concat(
+            [
+                pd.Series(repeat(values[i], len(period)), index=period)
+                for i, period in enumerate(periods_index)
+            ]
+        )
+
+        return periodic_values.tolist()
+
+    def create_yearly_values(
+        values: typing.Iterable[float], period_years: typing.Iterable[int]
+    ):
+        """
+        Creates a value for every year (between two periods)
+        Value of period is continued until next period
+        Parameters
+        ----------
+        values values to be interpolated
+        years years of periods
+
+        Returns list
+        -------
+
+        """
+        results = pd.Series()
+        for i in range(len(period_years) - 1):
+            diff = period_years[i + 1] - period_years[i]
+            period_results = pd.Series(repeat(values[i], diff))
+            results = pd.concat([results, period_results])
+        results = pd.concat([results, pd.Series(values[-1])])
+        return results.tolist()
+
+    def unpack_sequences(facade, period_data):
+        """
+        Depending on dtype and content:
+        Periodically changing values [given as array] are either unpacked into
+            - full periods
+            - yearly values (between periods)
+            - kept as periodical values
+
+        Decision happens based on
+            - value
+            - name
+            - entry in yearly/periodical values list.
+
+        Parameters
+        ----------
+        facade
+        period_data
+
+        Returns
+        -------
+        facade
+        """
+
+        yearly_values = ["fixed_costs", "marginal_costs"]
+        periodical_values = [
+            "capacity",
+            "capacity_cost",
+            "capacity_potential",
+            "storage_capacity",
+        ]
+
+        for value_name, value in facade.items():
+            if isinstance(value, Decimal):
+                facade[value_name] = float(value)
+            # check if multi-period and value is list
+            if period_data and isinstance(value, list):
+                # check if length of list equals number of periods
+                if len(value) == len(period_data["periods"]):
+                    if value_name in periodical_values:
+                        # special period parameters don't need to be
+                        # converted into timeseries
+                        facade[value_name] = [
+                            float(vv) if isinstance(vv, Decimal) else vv
+                            for vv in value
+                        ]
+                        continue
+                    elif value_name in yearly_values:
+                        # special period parameter need to be
+                        # converted into timeseries with value for each
+                        # year
+                        facade[value_name] = create_yearly_values(
+                            value, period_data["years"]
+                        )
+                        msg = (
+                            f"\nThe parameter '{value_name}' of a "
+                            f"'{facade['type']}' facade is converted "
+                            "into a yearly list. This might not be "
+                            "possible for every parameter and lead to "
+                            "ambiguous error messages.\nPlease be "
+                            "aware, when using this feature!"
+                        )
+                        warnings.warn(msg, UserWarning)
+
+                    else:
+                        # create timeseries with periodic values
+                        facade[value_name] = create_periodic_values(
+                            value, period_data["periods"]
+                        )
+                        msg = (
+                            f"\nThe parameter '{value_name}' of a "
+                            f"'{facade['type']}' facade is converted "
+                            "into a periodic timeseries. This might "
+                            "not be possible for every parameter and "
+                            "lead to ambiguous error messages.\nPlease"
+                            " be aware, when using this feature!"
+                        )
+                        warnings.warn(msg, UserWarning)
+        return facade
+
     facades = {}
     for r in package.resources:
         if all(
@@ -400,13 +561,12 @@ def deserialize_energy_system(cls, path, typemap={}, attributemap={}):
                 fk["fields"]: fk["reference"]
                 for fk in r.descriptor["schema"].get("foreignKeys", ())
             }
+
             for facade in facade_data:
                 # convert decimal to float
-                for f, v in facade.items():
-                    if isinstance(v, Decimal):
-                        facade[f] = float(v)
+
                 read_facade(
-                    facade,
+                    unpack_sequences(facade=facade, period_data=period_data),
                     facades,
                     create,
                     typemap,
@@ -442,28 +602,12 @@ def deserialize_energy_system(cls, path, typemap={}, attributemap={}):
         # if no temporal provided as resource, take the first timeindex
         # from dict
         else:
-            # look for periods resource and if present, take as periods from it
+            # look for periods resource and if present, take periods from it
             if package.get_resource("periods"):
-                df_periods = pd.DataFrame.from_dict(
-                    package.get_resource("periods").read(keyed=True)
-                )
-                timeincrement = df_periods["increment"].values
-                timeindex = pd.DatetimeIndex(df_periods["timeindex"])
-                periods = [
-                    pd.DatetimeIndex(df["timeindex"])
-                    for period, df in df_periods.groupby("periods")
-                ]
-                periods = [
-                    pd.DatetimeIndex(
-                        i.values, freq=i.inferred_freq, name="timeindex"
-                    )
-                    for i in periods
-                ]
-
                 es = cls(
-                    timeindex=timeindex,
-                    timeincrement=timeincrement,
-                    periods=periods,
+                    timeindex=period_data["timeindex"],
+                    timeincrement=period_data["timeincrement"],
+                    periods=period_data["periods"],
                     infer_last_interval=False,
                 )
 
@@ -483,7 +627,7 @@ def deserialize_energy_system(cls, path, typemap={}, attributemap={}):
                 timeindex = pd.date_range(
                     start=pd.to_datetime("today"), periods=1, freq="H"
                 )
-                es = cls()
+                es = cls(timeindex=timeindex)
 
         es.add(
             *chain(