+++ src/oemof/tabular/examples/datapackages/dispatch_multi_period/scripts/infer.py
@@ -2,6 +2,7 @@
 Run this script from the root directory of the datapackage to update
 or create meta data.
 """
+
 from oemof.tabular.datapackage import building
 
 # This part is for testing only: It allows to pass